rename option to build corporate website
reason: be more explicit and avoid confusion
