+++ src/component/Layout.js
@@ -1,4 +1,4 @@
-import { forDeveloper } from '../../config-wazo';
+import { corporate } from '../../config-wazo';
 
 import '../styles/bootstrap.css';
 import '../styles/fontawesome.css';
@@ -6,6 +6,6 @@ import '../styles/prism.css';
 import '../styles/elegant-font.css';
 import 'mainCSS/index.scss';
 
-const Layout = forDeveloper ? require('./LayoutDev').default : require('./LayoutPlatform').default;
+const Layout = corporate ? require('./LayoutCorporate').default : require('./LayoutPlatform').default;
 
 export default Layout;
