+++ src/component/corporate/index.js
