+++ src/html.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from "react"
 import PropTypes from "prop-types"
-import { forDeveloper } from '../config-wazo';
+import { corporate } from '../config-wazo';
 
 export default function HTML(props) {
   return (
@@ -14,7 +14,7 @@ export default function HTML(props) {
         />
         <link rel="icon"
           type="image/png"
-          href={`/images/${forDeveloper ? 'dev' : 'platform'}/icon.png`}
+          href={`/images/${corporate ? 'corporate' : 'platform'}/icon.png`}
         />
         {props.headComponents}
         <script type="text/javascript" src="https://code.jquery.com/jquery-3.2.1.min.js" />
